moved dto in another file
